// types.ts
export enum ViewState {
  PORTAL = 'PORTAL',
  PRE_ARCHIVE = 'PRE_ARCHIVE',
  COLLECTION = 'COLLECTION',
  ARCHIVE_MGMT = 'ARCHIVE_MGMT',
  QUERY = 'QUERY',
  BORROWING = 'BORROWING',
  WAREHOUSE = 'WAREHOUSE',
  STATS = 'STATS',
  SETTINGS = 'SETTINGS',
  ADMIN = 'ADMIN',
}
// ...
export type ColumnType = 'text' | 'status' | 'progress' | 'money' | 'date' | 'action';

export interface TableColumn {
  key: string;
  header: string;
  type: ColumnType;
  width?: string;
}

export interface GenericRow {
  id: string;
  [key: string]: any;
}

export interface ModuleConfig {
  columns: TableColumn[];
  data: GenericRow[];
}